+++ CHANGELOG.md
@@ -21,6 +21,7 @@ bugs in actually-released versions.
 - The webapp automatically reloads after a disconnect, when it reconnects to Flamenco Manager and sees the Manager version changed [#104235](https://projects.blender.org/studio/flamenco/pulls/104235).
 - Show the configured Flamenco Manager name in the webapp's browser window title.
 - Workers can be marked as 'restartable' by using the `-restart-exit-code N` commandline option. More info in the [Worker Actions documentation](https://flamenco.blender.org/usage/worker-actions/).
+- The `{timestamp}` placeholder in the render output path is now replaced with a local timestamp (rather than UTC).
 
 
 ## 3.2 - released 2023-02-21